requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.2
django-crispy-forms==1.8.1
django-filter==2.3.0
django-import-export==2.1.0
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.0
tablib==1.1.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
